==> reduce_analysis/2b_Weibull_REDUCE_Model_20220501.Rmd <==
---
title: "Reduce Analysis"
author: "Lacey Etzkorn"
date: "5/01/2021"
output: html_document
params: 
  output_dir: "../reduce_analysis_output/"
---

```{r setup, echo = FALSE,message = FALSE}
library(frailtypack); library(gt); library(tidyverse)
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
load(file = "../reduce_data/processed_data_coma_delirium_2mgonly.rdata")
df2 <- df2 %>% ungroup %>%
filter(study_arm!=3 & (apache > 0) & !is.na(apache))%>% 
mutate(gap = tstop - tstart,
       id = dense_rank(id),
       treatment = 1*(study_arm==2),
	   age60 = 1*(age<60),
	   age70 = 1*(age<70), 
	   age80 = 1*(age<80))
#table(df2$apache, useNA = "ifany")
#table(df2$age, useNA = "ifany")
```

# Treatment (2mg) Effect Only Models

```{r, echo = FALSE,message = FALSE}
if(!file.exists("../reduce_analysis_output/REDUCE_Weibull_Models_20220501.rdata")){

# the death initialization model for the gap.model doesn't work, so we set some starting vals
# gap.model1 <-
# frailtyPenal(formula = Surv(gap, del) ~ treatment+
# 		terminal(death) + cluster(id),
# 	formula.terminalEvent = ~treatment,
# 	data = df2, 
# 	recurrentAG = F,
# 	hazard = "Weibull", RandDist = "LogN", 
# 	init.Alpha = 14,
# 	init.B = c(-0.0182,-0.0504),
# 	init.Theta = 0.238^2,
# 	maxit = 300)

#############################
# GAP Time Models (with/without initialization)
gap.model <-
multivPenal(formula = Surv(tstart, tstop, del) ~ treatment +
		terminal(death) + terminal2(discharge) + cluster(id),
	formula.terminalEvent = ~treatment,
	formula.terminalEvent2= ~treatment,
	data = df2,
	gapTimes = T,
	jointGeneral = F,
	hazard = "Weibull",
	maxit = 300, 
	initialize = T)

# # Extract Transformed Parameters
# gap.model$initialization$joint1 <- gap.model1
#  f1 <- function(b){
#  	c(b[1:(length(b)-3+1-2)]^2,
#  	  b[(length(b)-3+1-1):length(b)])
#  }
#  f1.prime <- function(b){
#  	diag(c(2*b[1:(length(b)-3+1-2)],
#  	       rep(1,3-1+2)))
#  }
#  Parameters1 = c("Recurrent: Shape", "Recurrent: Scale",
#  	  "Terminal1: Shape", "Terminal1: Scale",
#  	  "Sigma", "Terminal1: Alpha",
#  	  paste0("Recurrent: ","treatment"),
#  	  paste0("Terminal1: ","treatment"))
#  gap.model$initialization$varH.Estimate1 <-
# 		f1.prime(gap.model$initialization$joint1$b) %*%
#  		gap.model$initialization$joint1$varHtotal %*%
#  		f1.prime(gap.model$initialization$joint1$b)
#  gap.model$initialization$summary.table1 <- tibble(
#  		Parameter = Parameters1,
#  		Raw = gap.model$initialization$joint1$b,
#  		Raw.SE = sqrt(diag(gap.model$initialization$joint1$varHtotal)),
#  		Estimate = f1(gap.model$initialization$joint1$b),
#  		Estimate.SE = sqrt(diag(gap.model$initialization$varH.Estimate1)),
#  		p = 2*pnorm(q = -abs(Raw), mean = 0, sd = Raw.SE)
#  )

#############################
# Calendar Time Model
cal.model <-
multivPenal(formula = Surv(tstart, tstop, del) ~ treatment +
		terminal(death) + terminal2(discharge) + cluster(id),
	formula.terminalEvent = ~treatment,
	formula.terminalEvent2= ~treatment,
	data = df2,
	jointGeneral = F,
	hazard = "Weibull",
	maxit = 300, 
	initialize = T)
save(gap.model, cal.model,
     file = "../reduce_analysis_output/REDUCE_Weibull_Models_20220501.rdata")
}else load("../reduce_analysis_output/REDUCE_Weibull_Models_20220501.rdata")
```

## Gap-Time Estimates

```{r}
m1 <-dplyr::select(cal.model$summary.table, Parameter, Estimate, Estimate.SE, p)
m2 <-dplyr::select(cal.model$initialization$summary.table1, Parameter, Estimate, Estimate.SE, p)
m3 <-dplyr::select(cal.model$initialization$summary.table2, Parameter, Estimate, Estimate.SE, p)
m4 <-dplyr::select(gap.model$summary.table, Parameter, Estimate, Estimate.SE, p)
m5 <-dplyr::select(gap.model$initialization$summary.table1, Parameter, Estimate, Estimate.SE, p)
m6 <-dplyr::select(gap.model$initialization$summary.table2, Parameter, Estimate, Estimate.SE, p)
```

```{r}
tab.start <- 
left_join(m4, m5, by = "Parameter") %>%
left_join(m6, by = "Parameter")
colnames(tab.start)<- c(" ",paste(rep(c("Fit", "(SE)", "p"),3),rep(1:3, each = 3)))
	

gt(tab.start , rowname_col = "Parameter")%>%
tab_stubhead(label = "Parameter")%>%
tab_header(
    title = md("**REDUCE Trial: Gap-Time Weibull Model Estimates**"))%>%
fmt_missing(columns = 1:10, missing_text = "")  %>%
text_transform(
  locations = cells_body(columns = c(3,6,9)),
  fn = function(x) ifelse(x == "","",paste0("(",x,")"))
  ) %>%
fmt_number(columns = c(2,3,5,6,8,9),decimals = 3)%>%
fmt_number(columns = c(4,7,10),decimals = 3) #%>%
#as_latex() %>%as.character() %>%cat()
```

## Calendar-Time Estimates

```{r}
tab.start <- 
left_join(m1, m2, by = "Parameter") %>%
left_join(m3, by = "Parameter")
colnames(tab.start)<- c(" ",paste(rep(c("Fit", "(SE)", "p"),3),rep(1:3, each = 3)))
	

gt(tab.start , rowname_col = "Parameter")%>%
tab_stubhead(label = "Parameter")%>%
tab_header(
    title = md("**REDUCE Trial: Calendar-Time Weibull Model Estimates**"))%>%
fmt_missing(columns = 1:10, missing_text = "")  %>%
text_transform(
  locations = cells_body(columns = c(3,6,9)),
  fn = function(x) ifelse(x == "","",paste0("(",x,")"))
  ) %>%
fmt_number(columns = c(2,3,5,6,8,9),decimals = 3)%>%
fmt_number(columns = c(4,7,10),decimals = 3) #%>%
#as_latex() %>%as.character() %>%cat()
```

# Include Age and Apache Score Effects

```{r, echo = FALSE,message = FALSE}
if(!file.exists("../reduce_analysis_output/REDUCE_Weibull_Models_Covariates_20220501.rdata")){
gap.model <-
multivPenal(formula = Surv(tstart, tstop, del) ~ I(apache/5 - 19/5) + I(age/10 - 65/10)+
		terminal(death) + terminal2(discharge) + cluster(id),
	formula.terminalEvent = ~ I(apache/5- 19/5)+ I(age/10 - 65/10),
	formula.terminalEvent2= ~ I(apache/5- 19/5)+ I(age/10 - 65/10),
	data = df2,
	gapTimes = T,
	jointGeneral = F,
	hazard = "Weibull",
	maxit = 300, 
	initialize = T)

cal.model <-
multivPenal(formula = Surv(tstart, tstop, del) ~ I(apache/5 - 19/5) + I(age/10 - 65/10)+
		terminal(death) + terminal2(discharge) + cluster(id),
	formula.terminalEvent = ~ I(apache/5- 19/5)+ I(age/10 - 65/10),
	formula.terminalEvent2= ~ I(apache/5- 19/5)+ I(age/10 - 65/10),
	data = df2,
	gapTimes = F,
	jointGeneral = F,
	hazard = "Weibull",
	maxit = 300, 
	initialize = T)
save(gap.model, cal.model, file = "../reduce_analysis_output/REDUCE_Weibull_Models_Covariates_20220501.rdata")
}else load("../reduce_analysis_output/REDUCE_Weibull_Models_Covariates_20220501.rdata")

gap.model$summary.table %>% 
dplyr::select(Parameter, Estimate, Estimate.SE, p) %>%
gt(rowname_col = "Parameter")%>%
tab_stubhead(label = "Parameter")%>%
tab_header(
    title = md("**REDUCE Trial: Gap-Time Weibull Model Estimates**"))%>%
fmt_missing(columns = 1:4, missing_text = "")  %>%
text_transform(
  locations = cells_body(columns = c(3)),
  fn = function(x) ifelse(x == "","",paste0("(",x,")"))
  ) %>%
fmt_number(columns = c(2:4),decimals = 3)
```

```{r}
cal.model$summary.table %>% 
dplyr::select(Parameter, Estimate, Estimate.SE, p) %>%
gt(rowname_col = "Parameter")%>%
tab_stubhead(label = "Parameter")%>%
tab_header(
    title = md("**REDUCE Trial: Calendar-Time Weibull Model Estimates**"))%>%
fmt_missing(columns = 1:4, missing_text = "")  %>%
text_transform(
  locations = cells_body(columns = c(3)),
  fn = function(x) ifelse(x == "","",paste0("(",x,")"))
  ) %>%
fmt_number(columns = c(2:4),decimals = 3)
```

# Additional Covariate Information

```{r}
df2 %>% group_by(id) %>% summarise(apache = apache[1], age = age[1]) %>% 
summarise(length(unique(id)),
		  mean(apache), mean(age), 
		  sd(apache), sd(age), 
		  cor(age, apache),
		  quantile(age,0.25),
		  quantile(age,0.5),
		  quantile(age,0.75),
		  mean(age<50), mean(age<60),mean(age<70), mean(age<80),
		  min(age), max(age)) %>%
as.list
```
