==> reduce_analysis/2e_Weibull_REDUCE_Model_ComaDeliriumSeparate_20220921.Rmd <==
---
title: "Reduce Analysis"
author: "Lacey Etzkorn"
date: "9/21/2021"
output: html_document
params: 
  output_dir: "../reduce_analysis_output/"
---

```{r setup, echo = FALSE,message = FALSE}
library(frailtypack); library(gt); library(tidyverse)
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
```

# Separate Models for Delirium and Coma

```{r, echo = FALSE,message = FALSE}
if(!file.exists("../reduce_analysis_output/REDUCE_Weibull_Models_Separate_Delirium_Coma_20220921.rdata")){
load(file = "../reduce_data/processed_data_coma_delirium_separate_2mg_only.rdata")
df3 <- df3 %>% 
ungroup %>%
filter((apache > 0) & !is.na(apache))%>% 
mutate(gap = tstop - tstart,
       treatment = 1*(study_arm==2)) %>%
as.data.frame	
	
######################
# GAP Time Models 
gap.model.coma <-
frailtyPenal(
	formula = Surv(gap, coma) ~ treatment +
		terminal(death) + cluster(id),
	formula.terminalEvent = ~treatment,
	data = df3,
	hazard = "Weibull") # did not converge

gap.model.delirium <-
frailtyPenal(
	formula = Surv(gap, delirium) ~ treatment +
		terminal(death) + cluster(id),
	formula.terminalEvent = ~treatment,
	data = df3,
	hazard = "Weibull")

#############################
# Calendar Time Model, Coma
cal.model.coma <-
frailtyPenal(
	formula = Surv(tstart, tstop, coma) ~ treatment +
		terminal(death) + cluster(id),
	formula.terminalEvent = ~treatment,
	data = df3,
	hazard = "Weibull")

cal.model.delirium <-
frailtyPenal(
	formula = Surv(tstart, tstop, delirium) ~ treatment +
		terminal(death) + cluster(id),
	formula.terminalEvent = ~treatment,
	data = df3,
	hazard = "Weibull")


save(gap.model.coma, gap.model.delirium, cal.model.coma, cal.model.delirium,
     file = "../reduce_analysis_output/REDUCE_Weibull_Models_Separate_Delirium_Coma_20220921.rdata")
}else load("../reduce_analysis_output/REDUCE_Weibull_Models_Separate_Delirium_Coma_20220921.rdata")
gap.model.coma
gap.model.delirium
cal.model.coma
cal.model.delirium
```
